Improved bookmark interactions
When opening a bookmark folder, all bookmarks will "collapse" to save space on the bookmarks bar. Hovering over a bookmark will expand it to reveal its destination.

Highlighting a bookmark folder for 3/4ths of a second will reveal ALL bookmark folder items (including items not in the highlighted folder). If the bookmark folder is the first in the bookmarks bar, then  highlighting it for an additional second will reveal ALL bookmarks in the bookmarks bar. When a folder is no longer highlighted, then all bookmarks will return to their collapsed state.

Closing all of the bookmark folders will return the bookmarks bar to the default "expanded" state.